+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Add `convert_textual_links` to `post_bake` recipe
+
 
 ## [v2.30.0] - 2026-01-05